# Conflicts:
#	src/main/java/soialNetworkApp/errors/ExceptionHandlerController.java
#	src/main/java/soialNetworkApp/repository/PersonsRepository.java
#	src/main/java/soialNetworkApp/service/CurrenciesService.java
#	src/main/java/soialNetworkApp/service/UsersService.java
#	src/main/java/soialNetworkApp/service/WeatherService.java
#	src/main/resources/db/changelog/liquibase-outputChangeLog_2.yaml
#	src/test/java/soialNetworkApp/service/WeatherServiceTest.java

+++ src/main/java/soialNetworkApp/api/response/DialogRs.java
@@ -16,7 +16,7 @@ public class DialogRs {
 
     @ApiModelProperty(value = "count of unread message", example = "15")
     @JsonProperty("unread_count")
-    private Integer unreadCount;
+    private Long unreadCount;
 
     @ApiModelProperty(value = "last message in dialog")
     @JsonProperty("last_message")
